# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mural Maz Lince API",
    "description": "API REST del sistema Mural Maz Lince. Digitaliza el tablero de anuncios estudiantil de la Universidad Autónoma de Occidente, Unidad Regional Mazatlán.\n\n**Estado:** Definitivo Sprint 10 — incluye pipeline de moderación (Vision + bad-words + dHash), panel admin y endpoints operativos de shadowban/cola.\n\n**Autenticación:** Cookies HttpOnly con JWT (ADR-04). Incluir `credentials: 'include'` en cada solicitud del cliente.\n\n**Referencia SRS:** v1.4 | **Referencia SAD:** v1.3 | **Referencia DDC:** v1.2",
    "version": "1.0.0-sprint10",
    "contact": {
      "name": "Miriam Alanis — Directora del Proyecto",
      "url": "https://github.com/devHamialco/MuralMazLince"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Entorno local (docker-compose)"
    },
    {
      "url": "https://muralmaz.railway.app",
      "description": "Producción (Railway.app) — disponible al cierre del Sprint 14"
    }
  ],
...
  "paths": {
...
    "/admin/moderation-queue": {
      "get": {
        "tags": ["admin"],
        "summary": "Cola de moderación",
        "description": "Lista todos los anuncios en estado `pending_review` y `shadowban`, ordenados: urgentes primero (`urgency_alert_at IS NOT NULL`), luego por fecha de creación. Incluye el motivo del marcado, datos del emprendedor y, para casos de hashing, la imagen existente para comparación visual. **RF-29, RF-37, RF-38.**",
        "operationId": "getModerationQueue",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Cola de moderación ordenada por urgencia.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ModerationQueueItem" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/admin/announcements/{id}/approve": {
      "patch": {
        "tags": ["admin"],
        "summary": "Aprobar anuncio pendiente",
        "description": "Cambia el estado del anuncio a `active`, lo publica en el feed y genera notificación interna al emprendedor. Registra `admin_id` y `resolved_at` en `moderation_queue`. **RF-30, RN-11.**",
        "operationId": "approveAnnouncement",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AnnouncementId" }
        ],
        "responses": {
          "200": {
            "description": "Anuncio aprobado y publicado en el feed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/admin/announcements/{id}/reject": {
      "post": {
        "tags": ["admin"],
        "summary": "Rechazar anuncio pendiente",
        "description": "Cambia el estado del anuncio a `rejected`, elimina su imagen de Cloudinary usando `cloudinary_id` y genera notificación interna al emprendedor con el motivo. **RF-30, RF-33, RN-11.**",
        "operationId": "rejectAnnouncement",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AnnouncementId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RejectAnnouncementRequest" },
              "example": { "reason": "La imagen contiene contenido inapropiado detectado por la moderación automática." }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Anuncio rechazado e imagen eliminada de Cloudinary.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "Anuncio rechazado. El emprendedor fue notificado." }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/admin/users/{id}/suspend": {
      "post": {
        "tags": ["admin"],
        "summary": "Suspender cuenta de usuario",
        "description": "Establece `is_suspended = true` en la cuenta del usuario indicado. El usuario suspendido recibe HTTP 403 al intentar iniciar sesión. **RF-30, RN-11.**",
        "operationId": "suspendUser",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UserId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SuspendUserRequest" },
              "example": { "reason": "Publicación reiterada de contenido inapropiado." }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta suspendida correctamente.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "Cuenta suspendida correctamente." }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/admin/claim-tickets": {
      "get": {
        "tags": ["admin"],
        "summary": "Listar tickets de reclamo de matrícula pendientes",
        "description": "Devuelve todos los tickets de reclamo con status `pending`, ordenados por fecha de creación ascendente. **RF-35.**",
        "operationId": "getClaimTickets",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de tickets pendientes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ClaimTicket" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/admin/claim-tickets/{id}/resolve": {
      "patch": {
        "tags": ["admin"],
        "summary": "Resolver ticket de reclamo",
        "description": "Marca el ticket como `resolved` y registra el `admin_id` y `resolved_at`. La acción concreta (reasignar o mantener la cuenta) queda fuera del sistema y se gestiona manualmente por el administrador vía WhatsApp. **RF-35.**",
        "operationId": "resolveClaimTicket",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ClaimTicketId" }
        ],
        "responses": {
          "200": {
            "description": "Ticket resuelto.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClaimTicket" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT en cookie HttpOnly. Emitido por POST /auth/login. No accesible desde JavaScript del cliente (ADR-04). Incluir `credentials: 'include'` en cada solicitud fetch del frontend."
      }
    },
...
    "parameters": {
      "AnnouncementId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID del anuncio.",
        "schema": { "type": "integer", "example": 42 }
      },
...
      "UserId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID del usuario.",
        "schema": { "type": "integer", "example": 3 }
      },
      "ClaimTicketId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID del ticket de reclamo.",
        "schema": { "type": "integer", "example": 2 }
      }
    },
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Descripción del error. Nunca expone datos sensibles." }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "code": { "type": "string", "example": "CAT-01" },
          "name": { "type": "string", "example": "Alimentos y bebidas" },
          "group_name": { "type": "string", "example": "Productos físicos" },
          "is_active": { "type": "boolean", "example": true }
        }
      },
      "AnnouncementSummary": {
        "type": "object",
        "description": "Vista resumida del anuncio para el feed. RF-08.",
        "properties": {
          "id": { "type": "integer", "example": 42 },
          "title": { "type": "string", "example": "Diseño de logos profesionales" },
          "cloudinary_url": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/mural/image/upload/v1/mural_lince/abc123.jpg" },
          "category": { "$ref": "#/components/schemas/Category" },
          "project_name": { "type": "string", "example": "Tech Lince — Diseño Gráfico" },
          "entrepreneur_display_name": {
            "type": "string",
            "description": "Proveniente de entrepreneur_profiles.display_name. RF-08, RF-39.",
            "example": "Mariana G. Tech Lince"
          },
          "total_likes": { "type": "integer", "example": 14 },
          "average_rating": { "type": "number", "format": "float", "example": 2.7 },
          "total_ratings": { "type": "integer", "example": 11 },
          "expires_at": { "type": "string", "format": "date-time", "example": "2026-05-15T23:59:59Z" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-10T14:30:00Z" }
        }
      },
      "AnnouncementDetail": {
        "type": "object",
        "description": "Vista completa del anuncio. Incluye contact_link solo para usuarios autenticados. RF-10, RF-11.",
        "allOf": [
          { "$ref": "#/components/schemas/AnnouncementSummary" },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "maxLength": 500,
                "example": "Diseño de logotipos, tarjetas de presentación y material visual para tu emprendimiento."
              },
              "status": {
                "type": "string",
                "enum": ["active", "pending_review", "rejected", "shadowban"],
                "example": "active"
              },
              "contact_link": {
                "type": "string",
                "format": "uri",
                "nullable": true,
                "description": "URL wa.me generada dinámicamente en el servidor. null para ROL-01. El número nunca se expone. RF-11, RN-09.",
                "example": "https://wa.me/5216691234567"
              },
              "project_id": { "type": "integer", "example": 7 },
              "custom_category": {
                "type": "string",
                "nullable": true,
                "description": "Solo presente si category.code = CAT-00 (Otro). RN-12.",
                "example": null
              }
            }
          }
        ]
      },
...
      "ModerationQueueItem": {
        "type": "object",
        "description": "Elemento de la cola de moderación para el administrador. RF-29.",
        "properties": {
          "moderation_id": { "type": "integer", "example": 5 },
          "announcement": { "$ref": "#/components/schemas/AnnouncementDetail" },
          "trigger_type": {
            "type": "string",
            "enum": ["vision_api", "bad_words", "phash", "report_threshold"],
            "example": "phash"
          },
          "trigger_detail": {
            "type": "object",
            "description": "JSON con el detalle del trigger. Estructura varía según trigger_type.",
            "example": { "hamming_distance": 7, "threshold": 10, "match_announcement_id": 38 }
          },
          "is_urgent": {
            "type": "boolean",
            "description": "true si urgency_alert_at IS NOT NULL. Indica anuncio con ≥3 reportes pendiente de acción. RF-37.",
            "example": true
          },
          "urgency_alert_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-04-12T22:00:00Z"
          },
          "report_count": {
            "type": "integer",
            "description": "Conteo de reportes únicos del anuncio. Relevante para trigger_type = report_threshold.",
            "example": 4
          },
          "entrepreneur": {
            "type": "object",
            "properties": {
              "user_id": { "type": "integer" },
              "matricula": { "type": "string" },
              "display_name": { "type": "string" }
            }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ClaimTicket": {
        "type": "object",
        "description": "Ticket de reclamo de matrícula. RF-35.",
        "properties": {
          "id": { "type": "integer", "example": 2 },
          "disputed_matricula": { "type": "string", "example": "20240001" },
          "status": { "type": "string", "enum": ["pending", "resolved"], "example": "pending" },
          "admin_id": { "type": "integer", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "resolved_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "RejectAnnouncementRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Motivo del rechazo. Se incluye en la notificación interna al emprendedor. RF-30.",
            "example": "La imagen contiene contenido inapropiado detectado por la moderación automática."
          }
        }
      },
      "SuspendUserRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Justificación de la suspensión. Queda registrada en el log de moderación. RF-30.",
            "example": "Publicación reiterada de contenido inapropiado."
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "400 Bad Request — Checksum inválido, campos faltantes o formato de WhatsApp incorrecto.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "El formato de la matrícula es inválido. Debe tener 8 dígitos y comenzar con 2." }
          }
        }
      },
      "Unauthorized": {
        "description": "401 Unauthorized — Cookie de sesión ausente o expirada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Sesión no encontrada. Por favor inicia sesión." }
          }
        }
      },
      "Forbidden": {
        "description": "403 Forbidden — Cuenta suspendida, o rol insuficiente para la acción solicitada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Tu cuenta está suspendida. Contacta al administrador." }
          }
        }
      },
      "NotFound": {
        "description": "404 Not Found — El recurso solicitado no existe.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Anuncio no encontrado." }
          }
        }
      },
...
      "InternalServerError": {
        "description": "500 Internal Server Error — Error inesperado en el servidor. El detalle se registra en el log; el cliente recibe un mensaje genérico.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Ocurrió un error interno. Por favor intenta más tarde." }
          }
        }
      }
    }
  }
}